PullRequest: sulong/206

+++ projects/com.oracle.truffle.llvm.nodes/src/com/oracle/truffle/llvm/nodes/asm/LLVMAMD64SarNode.java
@@ -46,7 +46,7 @@ public abstract class LLVMAMD64SarNode extends LLVMExpressionNode {
     @NodeChildren({@NodeChild("left"), @NodeChild("right")})
     public abstract static class LLVMAMD64SarwNode extends LLVMExpressionNode {
         @Specialization
-        protected short executeI16(short left, short right) {
+        protected short executeI16(short left, byte right) {
             return (short) (left >> right);
         }
     }
@@ -54,7 +54,7 @@ public abstract class LLVMAMD64SarNode extends LLVMExpressionNode {
     @NodeChildren({@NodeChild("left"), @NodeChild("right")})
     public abstract static class LLVMAMD64SarlNode extends LLVMExpressionNode {
         @Specialization
-        protected int executeI32(int left, int right) {
+        protected int executeI32(int left, byte right) {
             return left >> right;
         }
     }
@@ -62,7 +62,7 @@ public abstract class LLVMAMD64SarNode extends LLVMExpressionNode {
     @NodeChildren({@NodeChild("left"), @NodeChild("right")})
     public abstract static class LLVMAMD64SarqNode extends LLVMExpressionNode {
         @Specialization
-        protected long executeI64(long left, long right) {
+        protected long executeI64(long left, byte right) {
             return left >> right;
         }
     }
